Moved AppSettings to Services
