+++ Controllers/UserController.cs
@@ -10,6 +10,7 @@ using Microsoft.IdentityModel.Tokens;
 using OOL_API.Data;
 using OOL_API.Models;
 using OOL_API.Models.DataTransfer;
+using OOL_API.Services;
 
 #nullable enable
